% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{birthplace_census}
\alias{birthplace_census}
\title{Census code and location name for Census birthplace locations}
\format{A data frame with 533 rows and 2 variables:
\describe{
  \item{code}{Census code for the birthplace}
  \item{location}{Name of the state, county, or region
   corresponding with the code}
  ...
}}
\source{
\url{https://usa.ipums.org/usa-action/variables/BPL#codes_section}
}
\usage{
birthplace_census
}
\description{
A dataset containing the code and location name for every
location included in the US Census birthplace list.
}
\keyword{datasets}

