% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{census_city}
\alias{census_city}
\title{Census code and city name for City geographies}
\format{A data frame with 1168 rows and 3 variables:
\describe{
  \item{Code}{Census code for the city}
  \item{City}{Name of the city corresponding with the code}
  \item{State}{Name of the state that the city is in}
  ...
}}
\source{
\url{https://usa.ipums.org/usa-action/variables/CITY#codes_section}
}
\usage{
census_city
}
\description{
A dataset containing the code and city name for every city included in the
US Census.
}
\keyword{datasets}

