% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{census_region}
\alias{census_region}
\title{Census code and region name for Census region geographies}
\format{A data frame with 17 rows and 2 variables:
\describe{
  \item{region}{Name of the region corresponding with the code}
  \item{code}{Census code for the region}
  ...
}}
\source{
\url{https://usa.ipums.org/usa-action/variables/REGION#codes_section}
}
\usage{
census_region
}
\description{
A dataset containing the code and region name for every region included in the
US Census region list.
}
\keyword{datasets}

