% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{stateFIP}
\alias{stateFIP}
\title{Census code and state name for Census State FIP geographies}
\format{A data frame with 62 rows and 2 variables:
\describe{
  \item{state}{Name of the state corresponding with the code}
  \item{FIP_code}{Census FIP code for the state}
  ...
}}
\source{
\url{https://usa.ipums.org/usa-action/variables/STATEFIP#codes_section}
}
\usage{
stateFIP
}
\description{
A dataset containing the code and state name for every state included in the
US Census State FIP list.
}
\keyword{datasets}

