% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{stateICP}
\alias{stateICP}
\title{Census code and state name for Census State ICP geographies}
\format{A data frame with 55 rows and 2 variables:
\describe{
  \item{state}{Name of the state corresponding with the code}
  \item{ICP_code}{Census ICP code for the state}
  ...
}}
\source{
\url{https://usa.ipums.org/usa-action/variables/STATEICP#codes_section}
}
\usage{
stateICP
}
\description{
A dataset containing the code and state name for every state included in the
US Census State ICP list.
}
\keyword{datasets}

